refactor(doctor): extract dm alias migration helpers
