+++ src/commands/doctor-legacy-config.ts
@@ -9,6 +9,7 @@ import {
   resolveTelegramPreviewStreamMode,
 } from "../config/discord-preview-streaming.js";
 import { DEFAULT_ACCOUNT_ID } from "../routing/session-key.js";
+import { normalizeDmAliases } from "./doctor-legacy-dm-aliases.js";
 
 export function normalizeCompatibilityConfigValues(cfg: OpenClawConfig): {
   config: OpenClawConfig;
@@ -19,87 +20,6 @@ export function normalizeCompatibilityConfigValues(cfg: OpenClawConfig): {
 
   const isRecord = (value: unknown): value is Record<string, unknown> =>
     Boolean(value) && typeof value === "object" && !Array.isArray(value);
-
-  const normalizeDmAliases = (params: {
-    provider: "slack" | "discord";
-    entry: Record<string, unknown>;
-    pathPrefix: string;
-  }): { entry: Record<string, unknown>; changed: boolean } => {
-    let changed = false;
-    let updated: Record<string, unknown> = params.entry;
-    const rawDm = updated.dm;
-    const dm = isRecord(rawDm) ? structuredClone(rawDm) : null;
-    let dmChanged = false;
-
-    const allowFromEqual = (a: unknown, b: unknown): boolean => {
-      if (!Array.isArray(a) || !Array.isArray(b)) {
-        return false;
-      }
-      const na = a.map((v) => String(v).trim()).filter(Boolean);
-      const nb = b.map((v) => String(v).trim()).filter(Boolean);
-      if (na.length !== nb.length) {
-        return false;
-      }
-      return na.every((v, i) => v === nb[i]);
-    };
-
-    const topDmPolicy = updated.dmPolicy;
-    const legacyDmPolicy = dm?.policy;
-    if (topDmPolicy === undefined && legacyDmPolicy !== undefined) {
-      updated = { ...updated, dmPolicy: legacyDmPolicy };
-      changed = true;
-      if (dm) {
-        delete dm.policy;
-        dmChanged = true;
-      }
-      changes.push(`Moved ${params.pathPrefix}.dm.policy → ${params.pathPrefix}.dmPolicy.`);
-    } else if (topDmPolicy !== undefined && legacyDmPolicy !== undefined) {
-      if (topDmPolicy === legacyDmPolicy) {
-        if (dm) {
-          delete dm.policy;
-          dmChanged = true;
-          changes.push(`Removed ${params.pathPrefix}.dm.policy (dmPolicy already set).`);
-        }
-      }
-    }
-
-    const topAllowFrom = updated.allowFrom;
-    const legacyAllowFrom = dm?.allowFrom;
-    if (topAllowFrom === undefined && legacyAllowFrom !== undefined) {
-      updated = { ...updated, allowFrom: legacyAllowFrom };
-      changed = true;
-      if (dm) {
-        delete dm.allowFrom;
-        dmChanged = true;
-      }
-      changes.push(`Moved ${params.pathPrefix}.dm.allowFrom → ${params.pathPrefix}.allowFrom.`);
-    } else if (topAllowFrom !== undefined && legacyAllowFrom !== undefined) {
-      if (allowFromEqual(topAllowFrom, legacyAllowFrom)) {
-        if (dm) {
-          delete dm.allowFrom;
-          dmChanged = true;
-          changes.push(`Removed ${params.pathPrefix}.dm.allowFrom (allowFrom already set).`);
-        }
-      }
-    }
-
-    if (dm && isRecord(rawDm) && dmChanged) {
-      const keys = Object.keys(dm);
-      if (keys.length === 0) {
-        if (updated.dm !== undefined) {
-          const { dm: _ignored, ...rest } = updated;
-          updated = rest;
-          changed = true;
-          changes.push(`Removed empty ${params.pathPrefix}.dm after migration.`);
-        }
-      } else {
-        updated = { ...updated, dm };
-        changed = true;
-      }
-    }
-
-    return { entry: updated, changed };
-  };
 
   const normalizePreviewStreamingAliases = (params: {
     entry: Record<string, unknown>;
@@ -228,12 +148,12 @@ export function normalizeCompatibilityConfigValues(cfg: OpenClawConfig): {
     let changed = false;
     if (provider !== "telegram") {
       const base = normalizeDmAliases({
-        provider,
         entry: rawEntry,
         pathPrefix: `channels.${provider}`,
       });
       updated = base.entry;
       changed = base.changed;
+      changes.push(...base.changes);
     }
     const providerStreaming = normalizeStreamingAliasesForProvider({
       provider,
@@ -255,12 +175,12 @@ export function normalizeCompatibilityConfigValues(cfg: OpenClawConfig): {
         let accountChanged = false;
         if (provider !== "telegram") {
           const res = normalizeDmAliases({
-            provider,
             entry: rawAccount,
             pathPrefix: `channels.${provider}.accounts.${accountId}`,
           });
           accountEntry = res.entry;
           accountChanged = res.changed;
+          changes.push(...res.changes);
         }
         const accountStreaming = normalizeStreamingAliasesForProvider({
           provider,
